requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Machine Learning & Anomaly Detection
pyod==1.1.3
tensorflow==2.15.0
keras==2.15.0

# Time Series Forecasting
prophet==1.1.5
statsmodels==0.14.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# API & Web
requests==2.31.0
streamlit==1.29.0
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# Airflow
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
faker==22.0.0
joblib==1.3.2
colorama==0.4.6

# Reporting
fpdf==1.7.2
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0